backend/requirements/base.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.3,<4.4
django-debug-toolbar>=4.2,<4.3
pytest>=7.4,<7.5
pytest-django>=4.7,<4.8
pytest-cov>=4.1,<4.2
pytest-mock>=3.12,<3.13
pytest-xdist>=3.5,<3.6
Faker>=20.1,<20.2
factory-boy>=3.3,<3.4
pytz>=2024.2 